Setting the map extent to the selected phase. Removing redundant comment
